Manejar error al borrar collection de typesense

+++ airbyte-integrations/connectors/destination-typesense/destination_typesense/destination.py
@@ -35,16 +35,18 @@ class DestinationTypesense(Destination):
             stream_name = configured_stream.stream.name
             if configured_stream.destination_sync_mode == DestinationSyncMode.overwrite:
                 try:
-                    logger.info(f"Borrando de typesense la collection {stream_name}")
-                    client.collections[stream_name].delete()
+                    collections = [collection['name'] for collection in client.collections.retrieve()]
+                    if stream_name in collections:
+                        logger.info(f"Borrando de typesense la collection {stream_name}")
+                        client.collections[stream_name].delete()
                     logger.info(f"Clonando collection template de typesense")
                     client.api_call.post('/collections?src_name=template_collection', {
                         "name": stream_name
                     })
                     logger.info(f"Se creó la collection {stream_name} exitosamente")
                 except Exception as e:
-                    logger.error(f"Error deleting previous records for typesense collection {stream_name}: {e}")
-                    pass
+                    logger.error(f"Error recreando collection de typesense {stream_name}: {e}")
+                    return
 
             writer = TypesenseWriter(client, stream_name, config.get("batch_size"))
             for message in input_messages:
